add google cartographer to this mess

+++ src/main/java/org/brigero/loading/util/OnUnreadabilityDataReceive.java
@@ -1,6 +1,6 @@
 package org.brigero.loading.util;
 
 public interface OnUnreadabilityDataReceive {
-    void onPointCloudData(double[][] cloud, double timeStamp);
-    void onIMUData(double[] angular_velocity, double[] linear_acceleration, double[] quaternion, double timeStamp);
+    void onPointCloudData(float[][] cloud, double timeStamp);
+    void onIMUData(float[] angular_velocity, float[] linear_acceleration, float[] quaternion, double timeStamp);
 }
